requirements.txt
# AI Providers
anthropic>=0.40.0
groq>=0.11.0

# Job Scraping
python-jobspy>=1.1.0

# Browser Automation
playwright>=1.45.0

# Telegram Bot
python-telegram-bot>=21.0

# Scheduling
apscheduler>=3.10.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0

# PDF Generation
reportlab>=4.0

# HTTP Client
httpx>=0.27.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Google Sheets (optional)
gspread>=6.0

# Encryption (Fernet for secrets at rest)
cryptography>=42.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
